backend/requirements.txt
asgiref==3.11.1
Django==5.2.11
djangorestframework==3.16.1
psycopg2-binary==2.9.11
sqlparse==0.5.5
tzdata==2025.3
django-cors-headers==4.4.0
scikit-learn==1.5.2
numpy==1.26.4
